requirements.txt
Pillow==10.1.0
numpy==1.26.4
opencv-python==4.10.0.82